Change key from ID to UID
The site id is DB data. It would be preferable to not reference identifiers managed by the DB in the project.yaml config file because this creates the possibility of referencing data that does not exist or that is different (a different DB may associate a different site id with the same logical site depending on creation order). This is not really a problem when there is only one DB, but in situations where there is separate test site DBs or local dev DBs it is best to keep a distinction between config that is shared in the code repos and data in the DB.

The project.yaml file references sites by the site uid, so it would be better to store that as part of the field config instead.

This will be a breaking change for existing users of the plugin.